+++ src/test/java/stroom/dashboard/expression/v1/TestExpressionParser.java
@@ -483,6 +483,17 @@ class TestExpressionParser {
         assertThat(out.toString()).isEqualTo("this is it");
     }
 
+
+    @Test
+    public void testConcat1Plus() throws ParseException {
+        final Generator gen = createGenerator("'this'+' is '+'it'");
+
+        gen.set(getVal(3D));
+
+        final Val out = gen.eval();
+        Assert.assertEquals("this is it", out.toString());
+    }
+
     @Test
     void testConcat2() throws ParseException {
         final Generator gen = createGenerator("concat(${val}, ' is ', 'it')");